move logic of getDocuments to firestore-methods.ts file
