+++ server/firestore-methods.ts
@@ -1,5 +1,24 @@
-import { doc, Firestore } from "@firebase/firestore";
+import { collection, doc, Firestore, getDocs, query } from "@firebase/firestore";
 import { getDoc } from "firebase/firestore";
+import { Showing } from "../types";
+
+export async function getAllShowings(database: Firestore){
+    const showingsCollection = collection(database, "showings")  
+    const showingsQuery = query(showingsCollection)
+    const snapshot = await getDocs(showingsQuery)
+
+    const fetchedShowings: Showing[] = []
+    snapshot.forEach((document: any)=>{
+        const showing = {
+            id: document.id,
+            ...document.data()
+        }
+
+        fetchedShowings.push(showing)
+    })
+
+    return fetchedShowings
+}
 
 export async function getShowingDetails(database: Firestore, showingId: string | undefined) {
     if (showingId === undefined) {
